+++ src/pages/CODAE/KitLancheRelatorioCodaePage.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Relatorio from "../../components/TourRequest/CODAE/Relatorio";
+import Relatorio from "../../components/SolicitacaoDeKitLanche/CODAE/Relatorio";
 import Page from "../../components/Shareable/Page/Page";
 
 export default props => (
