+++ src/pages/DRE/KitLancheOrdersPage.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Page from "../../components/Shareable/Page/Page";
-import OrdersDashboardContainer from "../../components/TourRequest/DRE/OrdersDashboardContainer";
+import OrdersDashboardContainer from "../../components/SolicitacaoDeKitLanche/CODAE/OrdersDashboardContainer";
 
 export default props => (
   <Page tituloRastro="solicitações">
