+++ src/pages/DRE/KitLancheRelatorioPage.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Relatorio from "../../components/TourRequest/DRE/Relatorio";
+import Relatorio from "../../components/SolicitacaoDeKitLanche/DRE/Relatorio";
 import Page from "../../components/Shareable/Page/Page";
 
 export default props => (
