+++ src/pages/Terceirizada/KitLancheOrdersTerceirizadaPage.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Page from "../../components/Shareable/Page/Page";
-import OrdersDashboardContainer from "../../components/TourRequest/Terceirizada/OrdersDashboardContainer";
+import OrdersDashboardContainer from "../../components/SolicitacaoDeKitLanche/Terceirizada/OrdersDashboardContainer";
 
 export default props => (
   <Page tituloRastro="solicitações">
